libcsm/os.py
#
#  MIT License
#
#  (C) Copyright 2023 Hewlett Packard Enterprise Development LP
#
#  Permission is hereby granted, free of charge, to any person obtaining a
#  copy of this software and associated documentation files (the "Software"),
#  to deal in the Software without restriction, including without limitation
#  the rights to use, copy, modify, merge, publish, distribute, sublicense,
#  and/or sell copies of the Software, and to permit persons to whom the
#  Software is furnished to do so, subject to the following conditions:
#
#  The above copyright notice and this permission notice shall be included
#  in all copies or substantial portions of the Software.
#
#  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
#  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
#  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
#  THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR
#  OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
#  ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
#  OTHER DEALINGS IN THE SOFTWARE.
#
"""
Module for handling command line calls.

.. note::
   The ``_CLI`` object is private, the intended usage is to use
   ``run_command``.

"""
from contextlib import contextmanager
from time import time
from subprocess import PIPE
from subprocess import Popen
import os

from libcsm.logger import Logger

LOG = Logger(__file__)


class _CLI:
    """
    An object to abstract the return result from ``run_command``.

    """
    _stdout = ''
    _stderr = ''
    _return_code = None
    _duration = None

    def __init__(self, args: [str, list], shell: bool = False) -> None:
        """
        If shell==True then the arguments will be converted to a string if
        a list was passed.
        The conversion is recommended by Popen's documentation:
            https://docs.python.org/3/library/subprocess.html

        :param args: The arguments (as a list or string) to run with Popen.
        :param shell: Whether to run Popen in a shell (default: False)
        """
        if shell and isinstance(args, list):
            self.args = ' '.join(args)
        else:
            self.args = args
        self.shell = shell
        self._run()

    def _run(self) -> None:
        """
        Run the arguments and set the object's class variables with the
        results.
        """
        start_time = time()
        try:
            with Popen(
                self.args,
                stdout=PIPE,
                stderr=PIPE,
                shell=self.shell,
            ) as command:
                stdout, stderr = command.communicate()
        except IOError as error:
            self._stderr = error.strerror
            self._return_code = error.errno
            LOG.error('Could not find command for given args: %s', self.args)
        else:
            self._stdout = stdout.decode('utf8')
            self._stderr = stderr.decode('utf8')
            self._return_code = command.returncode
        self._duration = time() - start_time
        if self._return_code and self._duration:
            LOG.info(
                '%s ran for %f (sec) with return code %i',
                self.args,
                self._duration,
                self._return_code
            )

    @property
    def stdout(self) -> str:
        """
        ``stdout`` from the command.
        """
        return self._stdout

    @property
    def stderr(self) -> str:
        """
        ``stderr`` from the command.
        """
        return self._stderr

    @property
    def return_code(self) -> str:
        """
        return code from the command.
        """
        return self._return_code

    @property
    def duration(self) -> float:
        """
        The duration of the running command.
        """
        return self._duration


@contextmanager
def chdir(directory: str, create: bool = False) -> None:
    """
    Changes into a given directory and returns to the original directory on
    exit.

    .. note::
       This does not wrap the yield in a 'try', everything done within
       the else is the user's responsibility.

    .. code-block:: python

        from libcsm.os import chdir

        print('doing things in /dir/foo')
        with chdir('/some/other/dir'):
            # doing things in /some/other/dir
        print('doing more things in /dir/foo')

    :param directory: Where you want to go.
    :param create: Whether you want the entire tree created or not.
    """
    original = os.getcwd()
    if not os.path.exists(directory) and create:
        os.makedirs(directory)
    try:
        os.chdir(directory)
    except OSError:
        LOG.warning('Invalid directory [%s]', directory)
    else:
        yield
    finally:
        os.chdir(original)


def run_command(
    args: [list, str],
    in_shell: bool = False,
    silence: bool = False, ) -> _CLI:
    """
    Runs a given command or list of commands by instantiating a ``CLI`` object.

    .. code-block:: python

        from libcsm.os import run_command

        result = run_command(['my', 'args'])
        print(vars(result))

    :param args: List of arguments to run, can also be a string. If a string,
    :param in_shell: Whether to use a shell when invoking the command.
    :param silence: Tells this not to output the command to console.
    """
    if not silence:
        LOG.info(
            'Running sub-command: %s (in shell: %s)', ' '.join(args), in_shell
        )
    return _CLI(args, shell=in_shell)
